+++ tests/__init__.py
@@ -1017,6 +1017,21 @@ FAKE_UPCOMING_INVOICE = InvoiceDict(
         "currency": "usd",
         "customer": FAKE_CUSTOMER["id"],
         "description": None,
+        "default_tax_rates": [
+            {
+                "id": "txr_fakefakefakefakefake0001",
+                "object": "tax_rate",
+                "active": True,
+                "created": 1570921289,
+                "description": None,
+                "display_name": "VAT",
+                "inclusive": True,
+                "jurisdiction": "Example1",
+                "livemode": False,
+                "metadata": {"djstripe_test_fake_id": "txr_fakefakefakefakefake0001"},
+                "percentage": 15.0,
+            }
+        ],
         "discount": None,
         "due_date": None,
         "ending_balance": None,
@@ -1036,6 +1051,14 @@ FAKE_UPCOMING_INVOICE = InvoiceDict(
                     "proration": False,
                     "quantity": 1,
                     "subscription": None,
+                    "tax_amounts": [
+                        {
+                            "amount": 261,
+                            "inclusive": True,
+                            "tax_rate": "txr_fakefakefakefakefake0001",
+                        }
+                    ],
+                    "tax_rates": [],
                     "type": "subscription",
                 }
             ],
@@ -1055,11 +1078,23 @@ FAKE_UPCOMING_INVOICE = InvoiceDict(
         "statement_descriptor": None,
         "subscription": FAKE_SUBSCRIPTION["id"],
         "subtotal": 2000,
-        "tax": None,
+        "tax": 261,
         "tax_percent": None,
         "total": 2000,
+        "total_tax_amounts": [
+            {
+                "amount": 261,
+                "inclusive": True,
+                "tax_rate": "txr_fakefakefakefakefake0001",
+            }
+        ],
         "webhooks_delivered_at": 1439218870,
     }
+)
+
+FAKE_TAX_RATE_EXAMPLE_1_VAT = load_fixture("tax_rate_txr_fakefakefakefakefake0001.json")
+FAKE_TAX_RATE_EXAMPLE_2_SALES = load_fixture(
+    "tax_rate_txr_fakefakefakefakefake0002.json"
 )
 
 FAKE_INVOICEITEM = {
@@ -1099,6 +1134,29 @@ FAKE_INVOICEITEM_II = {
     "quantity": None,
     "subscription": None,
 }
+
+# Invoice item with tax_rates
+# TODO generate this
+FAKE_INVOICEITEM_III = {
+    "id": "ii_16XVTY2eZvKYlo2Cxz5n3RaS",
+    "object": "invoiceitem",
+    "amount": 2000,
+    "currency": "usd",
+    "customer": FAKE_CUSTOMER_II["id"],
+    "date": 1439033216,
+    "description": "One-time setup fee",
+    "discountable": True,
+    "invoice": FAKE_INVOICE_II["id"],
+    "livemode": False,
+    "metadata": {"key1": "value1", "key2": "value2"},
+    "period": {"start": 1439033216, "end": 1439033216},
+    "plan": None,
+    "proration": False,
+    "quantity": None,
+    "subscription": None,
+    "tax_rates": [FAKE_TAX_RATE_EXAMPLE_1_VAT],
+}
+
 
 FAKE_TRANSFER = {
     "id": "tr_16Y9BK2eZvKYlo2CR0ySu1BA",